+++ src/collective/dms/mailcontent/dmsmail.py
@@ -93,7 +93,7 @@ class IDmsIncomingMail(IDmsDocument):
     reply_to = RelatedDocs(
         title=_(u"In Reply To"),
         required=False,
-        portal_types=('dmsoutgoingmail',))
+        portal_types=('dmsincomingmail', 'dmsoutgoingmail'))
 
     form.order_before(sender='treating_groups')
     form.order_after(recipients='sender')
@@ -209,7 +209,7 @@ class IDmsOutgoingMail(IDmsDocument):
     reply_to = RelatedDocs(
         title=_(u"In Reply To"),
         required=False,
-        portal_types=('dmsincomingmail',))
+        portal_types=('dmsincomingmail', 'dmsoutgoingmail'))
 
     form.order_before(sender='treating_groups')
     form.order_before(recipients='treating_groups')