+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Add a recipients index, containing also the organisation chain UIDs [sgeulette]
 - Updated reply_to field to allow dmsincomingmail and dmsoutgoingmail types. [sgeulette]
 - Corrected indexer not working with 2 decorators. [sgeulette]
+- Added display_backrefs on reply_to field. [sgeulette]
 
 0.5 (2016-04-15)
 ----------------