+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog
 - Added display_backrefs on reply_to field. [sgeulette]
 - Combined title and internal_reference_no in Title for OutgoingMail [sgeulette]
 - Set internal_reference_no if empty [sgeulette]
+- Add internal_reference_no in SearchableText [sgeulette]
 
 0.5 (2016-04-15)
 ----------------